chore: Move `notifier` module
