+++ service/src/dependencies/mod.rs
@@ -3,6 +3,7 @@
 // Copyright: 2024, Rémi Bardon <remi@remibardon.name>
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
+pub mod any_notifier;
 mod notifier;
 mod uuid;
